software/server/requirements.txt
# 目前是直接手写的依赖，尚未验证版本是否有问题
dearpygui>=1.10.1
pyyaml>=6.0.1